+++ package.json
@@ -6,8 +6,8 @@
   "description": "Aaron Cordova - JavaScript Developer resume built with React!",
   "keywords": ["aaron", "cordova", "qodesmith", "resume", "react"],
   "scripts": {
-    "dev": "bun --hot src/index.ts",
-    "start": "NODE_ENV=production bun src/index.ts",
+    "dev": "bun --hot src/index.html",
+    "start": "NODE_ENV=production bun src/index.html",
     "build": "bun run build.ts",
     "check": "biome check",
     "check:fix": "biome check --write .",
