+++ src/App.tsx
@@ -1,49 +1,33 @@
-import {
-  Card,
-  CardContent,
-  CardDescription,
-  CardHeader,
-  CardTitle,
-} from '@/components/ui/card'
+/**
+ * This file is the entry point for the React app, it sets up the root
+ * element and renders the App component to the DOM.
+ *
+ * It is included in `src/index.html`.
+ */
 
-import {APITester} from './APITester'
-import './index.css'
+import {StrictMode} from 'react'
+import {createRoot} from 'react-dom/client'
 
-import logo from './logo.svg'
-import reactLogo from './react.svg'
+import {Resume} from './components/custom/Resume'
 
-export function App() {
-  return (
-    <div className="container relative z-10 mx-auto p-8 text-center">
-      <div className="mb-8 flex items-center justify-center gap-8">
-        <img
-          src={logo}
-          alt="Bun Logo"
-          className="h-36 scale-120 p-6 transition-all duration-300 hover:drop-shadow-[0_0_2em_#646cffaa]"
-        />
-        <img
-          src={reactLogo}
-          alt="React Logo"
-          className="h-36 p-6 transition-all duration-300 [animation:spin_20s_linear_infinite] hover:drop-shadow-[0_0_2em_#61dafbaa]"
-        />
-      </div>
-      <Card>
-        <CardHeader className="gap-4">
-          <CardTitle className="font-bold text-3xl">Bun + React</CardTitle>
-          <CardDescription>
-            Edit{' '}
-            <code className="rounded bg-muted px-[0.3rem] py-[0.2rem] font-mono">
-              src/App.tsx
-            </code>{' '}
-            and save to test HMR
-          </CardDescription>
-        </CardHeader>
-        <CardContent>
-          <APITester />
-        </CardContent>
-      </Card>
-    </div>
-  )
+import '../styles/globals.css'
+
+const elem = document.getElementById('root')
+
+if (!elem) throw new Error('No root element found')
+
+const app = (
+  <StrictMode>
+    <Resume />
+  </StrictMode>
+)
+
+if (import.meta.hot) {
+  // With hot module reloading, `import.meta.hot.data` is persisted.
+  // biome-ignore lint/suspicious/noAssignInExpressions: bun init created this
+  const root = (import.meta.hot.data.root ??= createRoot(elem))
+  root.render(app)
+} else {
+  // The hot module reloading API is not available in production.
+  createRoot(elem).render(app)
 }
-
-export default App
