+++ src/components/custom/Resume.tsx
@@ -1,0 +1,3 @@
+export function Resume() {
+  return <div className="h-[11in] w-[8.5in] bg-blue-500">Resume</div>
+}
