--- src/components/ui/label.tsx
@@ -1,25 +1,0 @@
-'use client'
-
-import type * as React from 'react'
-
-import {cn} from '@/lib/utils'
-
-import * as LabelPrimitive from '@radix-ui/react-label'
-
-function Label({
-  className,
-  ...props
-}: React.ComponentProps<typeof LabelPrimitive.Root>) {
-  return (
-    <LabelPrimitive.Root
-      data-slot="label"
-      className={cn(
-        'flex select-none items-center gap-2 font-medium text-sm leading-none peer-disabled:cursor-not-allowed peer-disabled:opacity-50 group-data-[disabled=true]:pointer-events-none group-data-[disabled=true]:opacity-50',
-        className
-      )}
-      {...props}
-    />
-  )
-}
-
-export {Label}
