+++ src/index.html
@@ -3,9 +3,9 @@
   <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="icon" type="image/svg+xml" href="./logo.svg">
-    <title>Bun + React</title>
-    <script type="module" src="./frontend.tsx" async></script>
+    <!-- <link rel="icon" type="image/svg+xml" href="./logo.svg"> -->
+    <title>Aaron Cordova</title>
+    <script type="module" src="./App.tsx" async></script>
   </head>
   <body>
     <div id="root"></div>
